+++ app/ferroterm-server/src/r4.rs
@@ -2,3 +2,5 @@
 //! over the engine (<https://hl7.org/fhir/R4/terminology-service.html>).
 
 crate::version::surface!(r4, "4.0.1", "R4", to_r4);
+
+crate::version::map::map!(r4);