+++ app/ferroterm-server/src/r4b.rs
@@ -2,3 +2,5 @@
 //! over the engine (<https://hl7.org/fhir/R4B/terminology-service.html>).
 
 crate::version::surface!(r4b, "4.3.0", "R4B", to_r4b);
+
+crate::version::map::map!(r4b);